# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type AdminUserApprovalStatus = typeof AdminUserApprovalStatus[keyof typeof AdminUserApprovalStatus];


export const AdminUserApprovalStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DECLINED: 'DECLINED',
} as const;

export type AdminUserSubscriptionPlan = typeof AdminUserSubscriptionPlan[keyof typeof AdminUserSubscriptionPlan];


export const AdminUserSubscriptionPlan = {
  NONE: 'NONE',
  ONE_DAY: 'ONE_DAY',
  TWO_DAYS: 'TWO_DAYS',
  THREE_DAYS: 'THREE_DAYS',
  LIFETIME: 'LIFETIME',
  CUSTOM: 'CUSTOM',
} as const;

export interface AdminUser {
  id: string;
  email: string;
  approvalStatus: AdminUserApprovalStatus;
  subscriptionPlan: AdminUserSubscriptionPlan;
  /**
     * @minimum 1
     * @maximum 365
     * @nullable
     */
  subscriptionDays: number | null;
  /** @nullable */
  subscriptionExpiresAt: string | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface AdminSession {
  admin: AdminUser;
}

export interface AdminSummary {
  total: number;
  pending: number;
  approved: number;
  declined: number;
}

export type SubscriptionApprovalInputPlan = typeof SubscriptionApprovalInputPlan[keyof typeof SubscriptionApprovalInputPlan];


export const SubscriptionApprovalInputPlan = {
  ONE_DAY: 'ONE_DAY',
  TWO_DAYS: 'TWO_DAYS',
  THREE_DAYS: 'THREE_DAYS',
  LIFETIME: 'LIFETIME',
  CUSTOM: 'CUSTOM',
} as const;

export interface SubscriptionApprovalInput {
  plan: SubscriptionApprovalInputPlan;
  /**
     * @minimum 1
     * @maximum 365
     * @nullable
     */
  customDays?: number | null;
}

export type ListAdminUsersParams = {
status?: ListAdminUsersStatus;
};

export type ListAdminUsersStatus = typeof ListAdminUsersStatus[keyof typeof ListAdminUsersStatus];


export const ListAdminUsersStatus = {
  ALL: 'ALL',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DECLINED: 'DECLINED',
} as const;