# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserApprovalStatus } from './adminUserApprovalStatus';
import type { AdminUserSubscriptionPlan } from './adminUserSubscriptionPlan';

export interface AdminUser {
  id: string;
  email: string;
  approvalStatus: AdminUserApprovalStatus;
  subscriptionPlan: AdminUserSubscriptionPlan;
  /**
     * @minimum 1
     * @maximum 365
     * @nullable
     */
  subscriptionDays: number | null;
  /** @nullable */
  subscriptionExpiresAt: Date | null;
  isAdmin: boolean;
  createdAt: Date;
}