# lib/api-zod/src/generated/types/subscriptionApprovalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionApprovalInputPlan } from './subscriptionApprovalInputPlan';

export interface SubscriptionApprovalInput {
  plan: SubscriptionApprovalInputPlan;
  /**
     * @minimum 1
     * @maximum 365
     * @nullable
     */
  customDays?: number | null;
}